infrastructure/requirements.txt
aws-cdk-lib==2.0.0
constructs>=10.0.0,<11.0.0

# Dot env (for load .env file)
python-dotenv==0.19.2
